notification_service/requirements.txt


pika==1.3.2
firebase-admin==6.4.0
python-dotenv==1.0.0